Refactor Stripe Connect handler.
