+++ src/ServiceProviders/Onboarding.php
@@ -4,9 +4,9 @@ namespace Give\ServiceProviders;
 
 use Give\Helpers\Hooks;
 use Give\Onboarding\Setup\Page as SetupPage;
-use Give\Onboarding\Setup\StripeConnectHandler;
 use Give\Onboarding\Wizard\Page as WizardPage;
 use Give\Onboarding\Setup\Handlers\TopLevelMenuRedirect;
+use Give\Onboarding\Setup\Handlers\StripeConnectHandler;
 
 class Onboarding implements ServiceProvider {
 
@@ -32,13 +32,10 @@ class Onboarding implements ServiceProvider {
 		// Maybe load Setup Page
 		if ( give_is_setting_enabled( SetupPage::getSetupPageEnabledOrDisabled() ) ) {
 			Hooks::addAction( 'admin_init', TopLevelMenuRedirect::class, 'maybeHandle' );
+			Hooks::addAction( 'admin_init', StripeConnectHandler::class, 'maybeHandle' );
 			Hooks::addAction( 'admin_menu', SetupPage::class, 'add_page' );
 			Hooks::addAction( 'admin_enqueue_scripts', SetupPage::class, 'enqueue_scripts' );
 			Hooks::addAction( 'admin_post_dismiss_setup_page', SetupPage::class, 'dismissSetupPage' );
 		}
-
-		// Handle Stripe Connect return.
-		// Priority 9 to listener implemented by the advanced settings.
-		Hooks::addAction( 'admin_init', StripeConnectHandler::class, 'maybeHandle', 9 );
 	}
 }
